// src/shared/types/socialMedia.ts
export enum SocialMediaPlatform {
  FACEBOOK = "facebook",
  INSTAGRAM = "instagram",
  TWITTER = "twitter",
  LINKEDIN = "linkedin",
  YOUTUBE = "youtube",
  TIKTOK = "tiktok",
  SNAPCHAT = "snapchat",
  PINTEREST = "pinterest",
  REDDIT = "reddit",
  DISCORD = "discord",
  TELEGRAM = "telegram",
  WHATSAPP = "whatsapp",
  GITHUB = "github",
  GITLAB = "gitlab",
  BEHANCE = "behance",
  DRIBBBLE = "dribbble",
  MEDIUM = "medium",
  DEV = "dev",
  HASHNODE = "hashnode",
  TWITCH = "twitch",
  SPOTIFY = "spotify",
  SOUNDCLOUD = "soundcloud",
  VIMEO = "vimeo",
  FLICKR = "flickr",
  TUMBLR = "tumblr",
}